requirements.txt
# AIVA CLI - Core Dependencies

# CLI Framework
typer==0.16.0

# Environment Management
python-dotenv==1.1.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# AI/ML Frameworks
crewai>=0.1.0
google-generativeai>=0.3.0

# Utilities
rich>=13.0.0
click>=8.0.0
pathlib2>=2.3.0

# Logging
structlog>=23.0.0
colorama>=0.4.6

# Development Dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0